# docs/swagger.yaml
definitions:
...
  model.Product:
    properties:
      airport_pickup:
        type: integer
      description:
        type: string
      id:
        type: integer
      id_travel_agent:
        type: integer
      lodging:
        type: integer
      name:
        type: string
      price:
        type: integer
      refund:
        type: integer
      tax:
        type: integer
    type: object
...
  model.TravelAgent:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      phone_number:
        type: string
    type: object
  model.TravelAgentSocialMedia:
    properties:
      facebook:
        type: string
      id:
        type: integer
      id_travel_agent:
        type: integer
      instagram:
        type: string
      tiktok:
        type: string
      twitter:
        type: string
      website:
        type: string
      youtube:
        type: string
    type: object
...
host: localhost:8888
info:
  contact: {}
  description: This is Travel Agent System API
  title: Travel Agent System API Documentation
  version: "2.0"
paths:
  /agent:
    get:
      consumes:
      - application/json
      description: User can get all agents information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TravelAgent'
      security:
      - JWT: []
      summary: Get All Travel Agents Information
      tags:
      - TravelAgent
    post:
      consumes:
      - application/json
      description: User can create agent
      parameters:
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.TravelAgent'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.TravelAgent'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TravelAgent'
      security:
      - JWT: []
      summary: Create Travel Agent
      tags:
      - TravelAgent
  /agent/{id}:
    delete:
      consumes:
      - application/json
      description: User can delete agent information if they want it
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TravelAgent'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TravelAgent'
      security:
      - JWT: []
      summary: Delete Travel Agent Information
      tags:
      - TravelAgent
    get:
      consumes:
      - application/json
      description: User can get agent information by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TravelAgent'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.TravelAgent'
      security:
      - JWT: []
      summary: Get Travel Agent Information by Id
      tags:
      - TravelAgent
    put:
      consumes:
      - application/json
      description: User can update agent information
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.TravelAgent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TravelAgent'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TravelAgent'
      security:
      - JWT: []
      summary: Update Travel Agent Information
      tags:
      - TravelAgent
...
  /product:
    get:
      consumes:
      - application/json
      description: User can get all products information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
      security:
      - JWT: []
      summary: Get All Products Information
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: User can create product
      parameters:
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Product'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Product'
      security:
      - JWT: []
      summary: Create Product
      tags:
      - Product
...
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: User can delete product information if they want it
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Product'
      security:
      - JWT: []
      summary: Delete Product Information
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: User can get product information by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Product'
      security:
      - JWT: []
      summary: Get Product Information by Id
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: User can update product information
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Product'
      security:
      - JWT: []
      summary: Update Product Information
      tags:
      - Product
...
  /travel-agent-social-media:
    get:
      consumes:
      - application/json
      description: User can get all travel agents social media information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TravelAgentSocialMedia'
      security:
      - JWT: []
      summary: Get All Travel Agents Social Media Information
      tags:
      - TravelAgentSocialMedia
    post:
      consumes:
      - application/json
      description: User can create travel agent social media information
      parameters:
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.TravelAgentSocialMedia'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.TravelAgentSocialMedia'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TravelAgentSocialMedia'
      security:
      - JWT: []
      summary: Create Travel Agent Social Media
      tags:
      - TravelAgentSocialMedia
  /travel-agent-social-media/{id}:
    delete:
      consumes:
      - application/json
      description: User can delete travel agent social media information if they want
        it
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TravelAgentSocialMedia'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TravelAgentSocialMedia'
      security:
      - JWT: []
      summary: Delete Travel Agent Social Media Information
      tags:
      - TravelAgentSocialMedia
    get:
      consumes:
      - application/json
      description: User can get travel agent social media information by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TravelAgentSocialMedia'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.TravelAgentSocialMedia'
      security:
      - JWT: []
      summary: Get Travel Agent Social Media Information by Id
      tags:
      - TravelAgentSocialMedia
    put:
      consumes:
      - application/json
      description: User can update travel agent social media information
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.TravelAgentSocialMedia'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TravelAgentSocialMedia'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TravelAgentSocialMedia'
      security:
      - JWT: []
      summary: Update Travel Agent Social Media Information
      tags:
      - TravelAgentSocialMedia
...
schemes:
- http
- https
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"